+++ Citadel/Changes.md
@@ -1,11 +1,15 @@
 # Utopia Disciple - List of changes
 
-## v1.0.1 (in development)
+## v1.1 (in development)
+### Breaking Changes
+- Changed sub-class and sub-object names in **DbOrganization**
 ### New Features
 - Added **DbTask** access point for accessing / viewing individual tasks' data
 ### New Methods
 - **DbUser**
   - Added `.email` and `.name` accessors to `.settings` -access point
+### Other Changes
+- **DbSingleOrganization** now contains `val organizationId`
 
 ## v1.0 - 13.7.2021
 Please refer to the README for more details.